Refactor harvester and prospector
Multiple small cleanups to harvester and prospector

* Move const Type to harvester package
* Rename harvester.Start to harvester.Run as this is consistent with the way we use Run
* Move log.go to harvester.go file for consistency
* Rename LogFile to File as it is now in the log package
* Rename fileReader to file and rename file to source in harvester
* Move json merging to reader from harvester
* Move source interface to harvester package and implementation into harvesters itself
* Move harvesters stats after init phase to make panic less likely (not preventing it)
* Create separate config namespace for log part as it is specific to the log object. This makes it possible to potentially reuse it in the future.
