+++ filebeat/harvester/reader/json.go
@@ -86,3 +86,22 @@ func (r *JSON) Next() (Message, error) {
 func createJSONError(message string) common.MapStr {
 	return common.MapStr{"message": message, "type": "json"}
 }
+
+// MergeJSONFields writes the JSON fields in the event map,
+// respecting the KeysUnderRoot and OverwriteKeys configuration options.
+// If MessageKey is defined, the Text value from the event always
+// takes precedence.
+func MergeJSONFields(data common.MapStr, jsonFields common.MapStr, text *string, config JSONConfig) {
+
+	// The message key might have been modified by multiline
+	if len(config.MessageKey) > 0 && text != nil {
+		jsonFields[config.MessageKey] = *text
+	}
+
+	if config.KeysUnderRoot {
+		// Delete existing json key
+		delete(data, "json")
+
+		jsontransform.WriteJSONKeys(data, jsonFields, config.OverwriteKeys)
+	}
+}
