+++ filebeat/harvester/source.go
@@ -1,17 +1,13 @@
-package source
+package harvester
 
 import (
 	"io"
 	"os"
 )
 
-type LogSource interface {
+type Source interface {
 	io.ReadCloser
 	Name() string
-}
-
-type FileSource interface {
-	LogSource
 	Stat() (os.FileInfo, error)
 	Continuable() bool // can we continue processing after EOF?
 	HasState() bool    // does this source have a state?
