+++ filebeat/prospector/log/file.go
@@ -1,4 +1,4 @@
-package source
+package log
 
 import "os"
